Applyes some DDD principles

+++ src/Services/Order/Order.Domain/Common/Enumeration.cs
@@ -9,9 +9,9 @@ namespace Order.Domain.Common
     {
         public string Name { get; private set; }
 
-        public int Id { get; private set; }
+        public short Id { get; private set; }
 
-        protected Enumeration(int id, string name)
+        protected Enumeration(short id, string name)
         {
             Id = id;
             Name = name;
